Improve the behaviour of zooming with Ctrl + scroll wheel
The previous behaviour was to recenter the view around the canvas position that was originally under the scroll wheel. This caused the image to shift around when using the scroll wheel at an off-center position.
Now, the logic is to keep the canvas position under the mouse the same, so that it's easy to Ctrl+scroll to zoom in on a particular area.
When zooming with +/- , the behaviour of zooming around the center of the screen is the same as before.

This needs some further testing, and likely the Zoom tool should also be updated to have similar behaviour.

Bug: 1958585